backend/core_service/requirements.txt
fastapi
uvicorn
httpx
pydantic
sqlalchemy
psycopg2
python-dotenv
pyjwt
loguru
